Create leaderboard form to write in it after game

+++ Assets/Scripts/Leaderboard/ILeaderboardService.cs
@@ -10,5 +10,5 @@ public interface ILeaderboardService
 [Serializable]
 public struct LeaderboardData {
     public string Name;
-    public int Seconds;
+    public float Seconds;
 }
